refactor(renderer): separate setup and workbench composition
