+++ apps/desktop/src/renderer/agents/connected-agents/ConnectedAgentsSettings.tsx
@@ -3,7 +3,7 @@ import { useEffect, useMemo, useRef, useState } from 'react'
 
 import type { ConnectedAgentSummary } from '../../../shared/contracts'
 import type { useConnectedAgents } from './useConnectedAgents'
-import { SettingsSection } from '../../components/SettingsSection'
+import { SettingsSection } from '../../app/settings/SettingsSection'
 
 function operationKey(label: string): string {
   return `${label}-${crypto.randomUUID()}`
