+++ apps/desktop/src/renderer/app/onboarding/OnboardingScreen.test.tsx
