+++ apps/desktop/src/renderer/app/settings/diagnostics/DiagnosticsPanel.test.tsx
