+++ apps/desktop/src/renderer/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 
-import { App } from './App'
+import { App } from './app/App'
 import './styles.css'
 import '@jobos/docx-editor-core/editor.css'
